--- TagsInput.css
@@ -1,73 +1,0 @@
-* {
-  font-family: "Roboto", sans-serif;
-  font-size: 16px;
-}
-
-body {
-  background: #f2f2f2;
-  padding: 20px;
-}
-
-.input-tag {
-  font-family: "Roboto", sans-serif;
-  font-size: 16px;
-  font-weight: 500;
-  color: #212121;
-  border: 1px solid #c3cfd9;
-  border-radius: 6px;
-  padding: 10px;
-  width: 180px;
-  margin: 0 20px 10px 0;
-  width: 100%;
-}
-
-.input-tag input {
-  border: none;
-  width: 100%;
-}
-
-.input-tag__tags {
-  display: inline-flex;
-  flex-wrap: wrap;
-  margin: 0;
-  padding: 0;
-  width: 100%;
-}
-
-.input-tag__tags li {
-  align-items: center;
-  background: #e9b75f;
-  border-radius: 6px;
-  color: white;
-  font-weight: bold;
-  font-size: 14px;
-  text-transform: uppercase;
-  display: flex;
-  list-style: none;
-  margin: 5px 5px 5px 0;
-  padding: 10px 15px;
-}
-
-.input-tag__tags li button {
-  align-items: center;
-  appearance: none;
-  background: #333333;
-  border: none;
-  border-radius: 50%;
-  color: white;
-  cursor: pointer;
-  display: inline-flex;
-  font-size: 12px;
-  height: 15px;
-  justify-content: center;
-  line-height: 0;
-  margin-left: 8px;
-  transform: rotate(45deg);
-  width: 15px;
-}
-
-.input-tag__tags li.input-tag__tags__input {
-  background: none;
-  flex-grow: 1;
-  padding: 0;
-}
