+++ src/components/templates/AddParticipantTeams.jsx
@@ -3,12 +3,11 @@ import styled from "styled-components";
 import { useSelector, useDispatch } from "react-redux";
 import { useHistory, useParams } from "react-router-dom";
 
-import { axiosWithAuth } from "../../utils/api";
 import { Footer } from "../organisms/index";
 import UserHeader from "../organisms/UserHeader";
 import WideBody from "../atoms/WideBody";
 import BodyContainer from "../atoms/BodyContainer";
-import { H3} from "../atoms/Heading";
+import { H3 } from "../atoms/Heading";
 import { RowHead } from "../atoms/RowHead";
 import { RowBody } from "../atoms/RowBody";
 import { Column } from "../atoms/Column";
@@ -16,39 +15,15 @@ import { CardWide } from "../atoms/Card";
 import Button from "../atoms/Button";
 import { type, Solid, media } from "../index";
 import { addParticipantTeamMember } from "../../store/participantTeams/actions";
+import { useSearchUserByEmail } from "../../hooks";
 
 const AddParticipantTeam = () => {
-  const [users, setUsers] = useState([]);
-  const [matches, setMatches] = useState([]);
-  const [searchString, setSearchString] = useState("");
   const [selectedUser, setSelectedUser] = useState(null);
   const dispatch = useDispatch();
   const history = useHistory();
   const { id } = useParams();
-  const { token } = useSelector(state => state.currentUser);
+  const [matches, searchString, setSearchString] = useSearchUserByEmail();
 
-  useEffect(() => {
-    const getUsers = async () => {
-      const {
-        data: {
-          body: { users }
-        }
-      } = await axiosWithAuth(token).get("/api/users");
-      setUsers(users);
-    };
-    getUsers();
-  }, [token]);
-
-  useEffect(() => {
-    const match = searchString
-      ? users
-          .filter(user =>
-            user?.email.toUpperCase().includes(searchString.toUpperCase())
-          )
-          .filter((_, i) => i < 5)
-      : [];
-    setMatches(match);
-  }, [searchString, users]);
 
   const { userId } = useSelector(state => state.currentUser);
   const createdTeam = useSelector(state =>
@@ -107,7 +82,7 @@ const AddParticipantTeam = () => {
         padding: 10px;
         margin: 0 20px 10px 0;
         ${({ display }) =>
-          display === "wide" ? `width: 100%;` : `width: 180px;`}
+        display === "wide" ? `width: 100%;` : `width: 180px;`}
 
         &:focus {
           transition: all 0.5s;
